backend/requirements.txt
Django==4.2.6
djangorestframework==3.15.1
drf_spectacular==0.27.2
python-dotenv==1.0.1
django-cors-headers
gunicorn
djangorestframework-simplejwt
django-extensions
django-filter
drf_spectacular
djangorestframework-jsonapi